Added all tests for UserResource.

+++ src/main/java/com/sky/identity/usermanagement/service/UserService.java
@@ -45,15 +45,15 @@ public class UserService {
 
     public UserDTO updateUserPassword(UpdatePasswordRequest request) {
         var user = userRepository.findById(request.getId()).orElseThrow(() -> new UserNotFoundException("User not found with id: " + request.getId()));
-        if (request.getPassword() == null || request.getPassword().isEmpty()) {
-            throw new IllegalArgumentException("Password must not be empty");
-        }
         user.setPassword(passwordEncoder.encode(request.getPassword()));
         userRepository.save(user);
         return new UserDTO(user.getId(), user.getEmail(), user.getUsername());
     }
 
     public void deleteUserById(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new UserNotFoundException("User not found with id: " + id);
+        }
         userRepository.deleteById(id);
     }
 }
